You can check how the changes were made in this commit: https://github.com/otg-br/otg-server/commit/128dde5a1a77cecaf79a67408801eeaa7dcd803e
Note: OTG Global and OTG Server share the same source base.

+++ engine/src/combat.cpp
@@ -18,6 +18,7 @@
  */
 
 #include "otpch.h"
+#include <cstdint>
 
 #include "combat.h"
 
@@ -421,7 +422,7 @@ bool Combat::setParam(CombatParam_t param, uint32_t value)
 		}
 
 		case COMBAT_PARAM_EFFECT: {
-			params.impactEffect = static_cast<uint8_t>(value);
+			params.impactEffect = static_cast<uint16_t>(value);
 			return true;
 		}
