+++ engine/src/combat.h
@@ -25,6 +25,7 @@
 #include "map.h"
 #include "baseevents.h"
 #include "monster.h"
+#include <cstdint>
 
 class Condition;
 class Creature;
@@ -74,7 +75,7 @@ struct CombatParams {
 	CombatType_t combatType = COMBAT_NONE;
 	CombatOrigin origin = ORIGIN_SPELL;
 
-	uint8_t impactEffect = CONST_ME_NONE;
+	uint16_t impactEffect = CONST_ME_NONE;
 	uint8_t distanceEffect = CONST_ANI_NONE;
 
 	bool blockedByArmor = false;
